fix : some functions moved to utils.py
